Moving common functions objects
