+++ services/worker/src/worker/common_exceptions.py
@@ -1,0 +1,18 @@
+from http import HTTPStatus
+from typing import Optional
+
+from worker.job_runner import JobRunnerError
+
+
+class StreamingRowsError(JobRunnerError):
+    """Raised when the rows could not be fetched in streaming mode."""
+
+    def __init__(self, message: str, cause: Optional[BaseException] = None):
+        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "StreamingRowsError", cause, True)
+
+
+class NormalRowsError(JobRunnerError):
+    """Raised when the rows could not be fetched in normal mode."""
+
+    def __init__(self, message: str, cause: Optional[BaseException] = None):
+        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "NormalRowsError", cause, True)
