+++ services/worker/src/worker/utils.py
@@ -1,10 +1,34 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 
-from typing import Any, List, Mapping, Optional, TypedDict
+import functools
+import itertools
+import logging
+import time
+import warnings
+from typing import (
+    Any,
+    Callable,
+    List,
+    Mapping,
+    Optional,
+    TypedDict,
+    TypeVar,
+    Union,
+    cast,
+)
 
-from datasets import Features
+from datasets import (
+    Dataset,
+    DatasetInfo,
+    DownloadConfig,
+    Features,
+    IterableDataset,
+    load_dataset,
+)
 from libcommon.utils import orjson_dumps
+
+from worker.common_exceptions import NormalRowsError, StreamingRowsError
 
 
 class DatasetItem(TypedDict):
@@ -204,3 +228,117 @@ def create_truncated_row_items(
             break
         row_items.append(row_item)
     return row_items
+
+
+FuncT = TypeVar("FuncT", bound=Callable[..., Any])
+
+
+def retry(func: FuncT) -> FuncT:
+    """retries with an increasing sleep before every attempt"""
+    SLEEPS = [1, 7, 70, 7 * 60, 70 * 60]
+    MAX_ATTEMPTS = len(SLEEPS)
+
+    @functools.wraps(func)
+    def decorator(*args: Any, **kwargs: Any) -> Any:
+        attempt = 0
+        last_err = None
+        while attempt < MAX_ATTEMPTS:
+            try:
+                """always sleep before calling the function. It will prevent rate limiting in the first place"""
+                duration = SLEEPS[attempt]
+                logging.info(f"Sleep during {duration} seconds to preventively mitigate rate limiting.")
+                time.sleep(duration)
+                return func(*args, **kwargs)
+            except ConnectionError as err:
+                logging.info("Got a ConnectionError, possibly due to rate limiting. Let's retry.")
+                last_err = err
+                attempt += 1
+        raise RuntimeError(f"Give up after {attempt} attempts with ConnectionError") from last_err
+
+    return cast(FuncT, decorator)
+
+
+@retry
+def get_rows(
+    dataset: str,
+    config: str,
+    split: str,
+    streaming: bool,
+    rows_max_number: int,
+    use_auth_token: Union[bool, str, None] = False,
+    column_names: Optional[List[str]] = None,
+) -> List[Row]:
+    download_config = DownloadConfig(delete_extracted=True)
+    ds = load_dataset(
+        dataset,
+        name=config,
+        split=split,
+        streaming=streaming,
+        use_auth_token=use_auth_token,
+        download_config=download_config,
+    )
+    if streaming:
+        if not isinstance(ds, IterableDataset):
+            raise TypeError("load_dataset should return an IterableDataset in streaming mode")
+    elif not isinstance(ds, Dataset):
+        raise TypeError("load_dataset should return a Dataset in normal mode")
+    if column_names:
+        ds = ds.select_columns(column_names)
+    rows_plus_one = list(itertools.islice(ds, rows_max_number + 1))
+    # ^^ to be able to detect if a split has exactly ROWS_MAX_NUMBER rows
+    if len(rows_plus_one) <= rows_max_number:
+        logging.debug(f"all the rows in the split have been fetched ({len(rows_plus_one)})")
+    else:
+        logging.debug(f"the rows in the split have been truncated ({rows_max_number} rows)")
+    return rows_plus_one[:rows_max_number]
+
+
+def get_rows_or_raise(
+    dataset: str,
+    config: str,
+    split: str,
+    rows_max_number: int,
+    use_auth_token: Union[bool, str, None],
+    info: DatasetInfo,
+    max_size_fallback: Optional[int] = None,
+    column_names: Optional[List[str]] = [],
+) -> List[Row]:
+    try:
+        return get_rows(
+            dataset=dataset,
+            config=config,
+            split=split,
+            streaming=True,
+            rows_max_number=rows_max_number,
+            use_auth_token=use_auth_token,
+            column_names=column_names,
+        )
+    except Exception as err:
+        MAX_SIZE_FALLBACK = 100_000_000
+        if max_size_fallback:
+            warnings.warn(
+                (
+                    f"The parameter 'max_size_fallback' is deprecated. The hard-coded value `{MAX_SIZE_FALLBACK}`"
+                    " will be used instead."
+                ),
+                category=DeprecationWarning,
+            )
+        if info.size_in_bytes is None or info.size_in_bytes > MAX_SIZE_FALLBACK:
+            raise StreamingRowsError(
+                "Cannot load the dataset split (in streaming mode) to extract the first rows.",
+                cause=err,
+            ) from err
+        try:
+            return get_rows(
+                dataset=dataset,
+                config=config,
+                split=split,
+                streaming=False,
+                rows_max_number=rows_max_number,
+                use_auth_token=use_auth_token,
+            )
+        except Exception as err:
+            raise NormalRowsError(
+                "Cannot load the dataset split (in normal download mode) to extract the first rows.",
+                cause=err,
+            ) from err
